build: pregen cgi/st, generated content in gen/
Pregenerating cgi/st results in two fewer new processes spun up per page
load. It also gets us closer to being distributable as a single
pre-built binary.

It will make development a little annoying, as it requires restarting
the server, but this issue should be tackled separately with a
development mode that hot-generates responses if the source repo is
available.