Build Werm as a self-contained binary
Now all cgi scripts and the logview util script are packed with the
binary.

We also now only write keyboard input to the subprocess when we know
the pty fd is ready for more writes. This may fix crash-like behavior
when pasting a large amount of text into a terminal.